+++ CHANGELOG.md
@@ -6,6 +6,11 @@ Change are listed in reverse chronological order (newest to oldest).
 
 <span class="changelog">
 
+###### [ 1.0.84 ] - 2024/08/18
+
+  * Updated `SpotifyClient.PlayerTransferPlayback` method to default the `refreshDeviceList` argument to True so that the Spotify Connect device list is refreshed by default before transferring playback.
+  * Updated `ZeroconfConnect` methods to use `connection:close` headers for requests to the Spotify Connect Zeroconf API endpoints.
+
 ###### [ 1.0.83 ] - 2024/08/15
 
   * Updated `SpotifyClient.GetSpotifyConnectDevice` method to check if a Spotify Desktop Client Application OAuth2 token exists when a PlayerTransfer targets a Sonos device.  If not found, then control will be transferred to the Sonos device using it's local queue; if found, then Spotify Connect will reconnect to the device and control will be transferred to the Sonos device using Spotify Connect.